requirements.txt
google-generativeai
google-genai
python-dotenv
requests
flask
gunicorn
google-cloud-firestore